split up StreamGroupSerializer* classes to new file. do the same for the tests
