+++ tests/sentry/api/serializers/test_group_stream.py
@@ -1,0 +1,42 @@
+from unittest import mock
+
+from sentry.api.serializers import serialize
+from sentry.api.serializers.models.group_stream import StreamGroupSerializer
+from sentry.models import Environment
+from sentry.testutils import TestCase
+
+
+class StreamGroupSerializerTestCase(TestCase):
+    def test_environment(self):
+        group = self.group
+
+        environment = Environment.get_or_create(group.project, "production")
+
+        from sentry.api.serializers.models.group import tsdb
+
+        with mock.patch(
+            "sentry.api.serializers.models.group.tsdb.get_range", side_effect=tsdb.get_range
+        ) as get_range:
+            serialize(
+                [group],
+                serializer=StreamGroupSerializer(
+                    environment_func=lambda: environment, stats_period="14d"
+                ),
+            )
+            assert get_range.call_count == 1
+            for args, kwargs in get_range.call_args_list:
+                assert kwargs["environment_ids"] == [environment.id]
+
+        def get_invalid_environment():
+            raise Environment.DoesNotExist()
+
+        with mock.patch(
+            "sentry.api.serializers.models.group.tsdb.make_series", side_effect=tsdb.make_series
+        ) as make_series:
+            serialize(
+                [group],
+                serializer=StreamGroupSerializer(
+                    environment_func=get_invalid_environment, stats_period="14d"
+                ),
+            )
+            assert make_series.call_count == 1
